test: consolidate 7 low-test-count specs into 3 to cut sandbox-compile overhead
Each HarnessSpec / ToolSpecBase / RuleHarnessSpec subclass pays a per-class
sandbox-compile cost (10-30s on Windows, per the timings in #183). Specs
with only 1-4 feature methods amortise that cost across very little
actual test work.

Merge seven thematically adjacent low-test-count specs into three
consolidated specs, eliminating four sandbox-compile passes from the
suite:

- server.ToolDeviceBasicsSpec  <- FindDeviceSpec + ToolGetDeviceSpec +
                                  ToolSendCommandSpec (11 tests, all
                                  exercise the device-tool primitives:
                                  findDevice resolver + toolGetDevice +
                                  toolSendCommand)

- server.ToolCustomRuleLifecycleSpec <- ToolCreateRuleSpec +
                                        ToolUpdateRuleSpec (6 tests,
                                        both test legacy custom_* rule
                                        lifecycle tools via TestChildApp
                                        Spy)

- rules.RuleEngineExecutionSpec <- RuleEngineSmokeSpec +
                                   ExecuteActionsSpec (4 tests, both
                                   test the rule child app's
                                   action-dispatch surface)

No production code touched. Test method names preserved verbatim
(lightly prefixed where one spec now covers more than one surface, e.g.
"toolGetDevice returns device summary shape...") so failures still
point at the underlying tool/method.

Verified locally: `./gradlew test` 1341 tests / 0 failures / 0 errors;
`python tests/sandbox_lint.py` clean.

Part of #183.
